modify deleteProperty.js

+++ app/actions/deleteProperty.js
@@ -10,27 +10,34 @@ async function deleteProperty(propertyId) {
   if (!sessionUser || !sessionUser.userId) {
     throw new Error("User ID is required");
   }
-  const { userId } = getSessionUser();
+  const { userId } = sessionUser;
   const property = await Property.findById(propertyId);
   if (!property) throw new Error("Property Not Found");
 
-  //Verify ownership
-  // if (property.owner.toString() !== userId) {
-  //   throw new Error("Unauthorized");
-  // }
-  //Extract public ID from image URLs
+  // Verify ownership
+  if (property.owner.toString() !== userId) {
+    throw new Error("Unauthorized");
+  }
+
+  // extract public id's from image url in DB
   const publicIds = property.images.map((imageUrl) => {
     const parts = imageUrl.split("/");
     return parts.at(-1).split(".").at(0);
   });
 
-  //Delete images from Cloudinary
+  // Delete images from Cloudinary
   if (publicIds.length > 0) {
     for (let publicId of publicIds) {
       await cloudinary.uploader.destroy("propertypulse/" + publicId);
     }
   }
+
+  // Proceed with property deletion
   await property.deleteOne();
+
+  // Revalidate the cache
+  // NOTE: since properties are pretty much on every page, we can simply
+  // revalidate everything that uses our top level layout
   revalidatePath("/", "layout");
 }
 export default deleteProperty;
